placed checks to make sure that long duration functions happen only one at a time

+++ app/survalytics/actions/survey.js
@@ -18,6 +18,10 @@ export function resetSkipSurvey() {
 
 export function deleteAllQuestions() {
     return (dispatch, getState) => {
+        if (getState().survalytic.status.deleteAllQuestions) {
+            return
+        }
+
         dispatch(deletingAllQuestions(true));
 
         Promise.all([DeleteAllQuestions(), DeleteAllResponses()])
@@ -25,6 +29,7 @@ export function deleteAllQuestions() {
                 dispatch(deletingAllQuestions(false));
             })
             .catch( (err) => {
+                dispatch(deletingAllQuestions(false));
                 console.log("ERROR (deleteAllQuestions): ", err);
             });
     };
@@ -33,6 +38,10 @@ export function deleteAllQuestions() {
 
 export function downloadSurvey(immediate = false) {
     return (dispatch, getState) => {
+        if (getState().survalytic.status.downloadingSurvey) {
+            return
+        }
+
         dispatch(downloadingSurvey(true));
 
         Download(immediate)
@@ -62,7 +71,11 @@ export function viewQuestions() {
 
 
 export function uploadResponses() {
-    return (dispatch) => {
+    return (dispatch, getState) => {
+        if(getState().survalytic.status.uploadingResponses) {
+            return
+        }
+        
         dispatch(uploadingResponses(true));
 
         Upload()
